new window view has been added

+++ src/components/DimentionsPage/DmgPage.jsx
@@ -6,9 +6,10 @@ import { BsCardText } from "react-icons/bs";
 import DimSetsList from "../Dimentions/DimSetsList";
 import BtnFontSize from "../EditBtns/BtnFontSize";
 import { constructDefItem, getNewOrParseDmg } from "../../constants/textParts";
-import { setRespNames } from "../../utils/localStorage";
 import DmgPageBody from "./DmgPageBody";
 import MyPortal from "../UI/MyPortal/MyPortal";
+import { sAlert } from "../../utils/alert";
+import ResponseFormatList from "../Dimentions/ResponseFormatList";
 
 const DmgPage = () => {
   const [isСheckerMode, setIsCheckerMode] = useState(false);
@@ -22,10 +23,19 @@ const DmgPage = () => {
     if (newVal !== action) setAction(newVal);
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, []);
-  const changeSet = (ns) => {
-    if (window.confirm("Change dimentions set? task will be cleared?"))
-      setItem(constructDefItem(ns));
+  const changeSet = async (ns) => {
+    const result = await sAlert({
+      title: "Change dimentions set? ",
+      text: "task will be cleared",
+      icon: "question",
+      showCancelButton: true,
+      confirmButtonText: "Yes, change it",
+      cancelButtonText: "Cancel",
+    });
+
+    if (result.isConfirmed) setItem(constructDefItem(ns));
   };
+  console.log(item);
 
   return (
     <div className={"bg-main pagedimwrap"} id="mainp">
@@ -33,14 +43,7 @@ const DmgPage = () => {
         <div className="bigmenu">
           <MyPortal containerId="navidPortal">
             <div className="small-menu-nav">
-              <button
-                // className="square-btn hotBtnGr rel-left"
-                className="btn-back m-0 round-btn"
-                onClick={() => {
-                  setRespNames(action, setAction);
-                }}>
-                {action}
-              </button>
+              <ResponseFormatList action={action} setAction={setAction} />
               <DimSetsList curSet={item.setName} setCurSet={changeSet} />
               <BtnFontSize />
               <Button
@@ -56,7 +59,7 @@ const DmgPage = () => {
           </MyPortal>
         </div>
 
-        <div className="txt-box pb-5">
+        <div className="txt-box">
           {isСheckerMode ? (
             <TextChecker close={() => setIsCheckerMode(!isСheckerMode)} />
           ) : (
